refactor(client/commands): qb-vehiclekeys:client:GiveKeys to separate file
